feat(root): update os.Exit value
fix(commit): print error message and exit

refactor(commit): remove unnecessary print statements

fix(pkg): update function name in utils package
